[iptables] Patch for simultaneous use of IPv4 and IPv6 rules. Fixes JB#44205
This commit adds a simple storage for loaded extension handles. The
storage can hold 128 handles (there are 122 extensions as total).
Both matches and targets that are loaded from extensions have their
handles stored.

When the address family is changed, all handles are closed with
dlclose(). Also the linked lists (xtables_pending_matches,
xtables_matches, xtables_pending_targets and xtables_targets are set to
NULL as the extensions registered to these have been unloaded.

This procedure solves the issue of not being able to add both IPv4 and
IPv6 rules within the same session from one program using xtables
library and socket options using ipt_replace/ip6t_replace. The reason
for this change is that _init() of the extension is only called when it
is loaded for the first time and for subsequent loads with dlopen() it
_init() is not executed.

Without this fix it results in a situation where IPv6 rules are added
first, the pending matches are loaded first and the library, e.g.,
multiport, is loaded with dlopen() and _init() registers the pending
matches for IPv6 family. When IPv4 rules are to be added from within the
same session after initializing with xtables_init_all() with new, changed
IP family, there is no match found for multiport as there are no pending
matches registered since call to _init() is not executed at shared library
load.

By unloading the extensions and nullifying the linked lists such
applications can add both IPv4 and IPv6 rules from within same session.